Criando a tabela comments, subordinada a post em relacionamento 1:n - scripts para a criação, comandos sql e inserção de commentários no banco efetivados -- precisou adequar os scripts sobre posts para eliminar a tabela subordinada comments antes de refazer a de posts

+++ create_post_tables.php
@@ -2,6 +2,13 @@
 
 $conn = require __DIR__.'/utils/connection.php';
 
+// apaga a tabela subordinada primeiro
+$sql = 'DROP TABLE comments';
+
+if(!$conn->query($sql)){
+	die('erro');
+}
+
 $sql = 'DROP TABLE posts';
 
 if(!$conn->query($sql)){
@@ -25,3 +32,5 @@ if(!$conn->query($sql)){
 $result = $conn->query('DESCRIBE posts');
 
 var_dump($result->fetch_all(MYSQLI_ASSOC));
+
+require __DIR__.'/create_comments_table.php';
